X-Wing/Swordfish: promote the row/column dispatch to a shared helper (#32)
The Swordfish de-dup centralized the row/column dispatch in a `line_of`
helper, but left it in swordfish.cpp's anonymous namespace -- so X-Wing, the
structural twin, still hand-rolled the identical idiom (an IIFE switch in the
find worker, and a runtime if/else collection in act_on_xwing). That is
exactly the "twins drift apart" the X-Wing/Swordfish pairing exists to avoid.

Promote the two helpers both fish were duplicating -- `line_of` (the Board
row/column mapping) and `candidates` (the note+value cell filter, previously
copy-pasted verbatim into each TU) -- into a shared analyzer-util.h, and route
both twins through them:

  - find_xwing: the `other_eset` IIFE switch collapses to one line_of call.
  - act_on_xwing: templatized on EliminationSet exactly like act_on_swordfish,
    so the two runtime branches collapse into one explicit-arg dispatch on
    is_row_based. CandidateSet and the elimination Unit are both derived from
    the single EliminationSet parameter (no mismatched <Row, Row> possible).
  - both fish drop their per-TU candidates()/line_of copies and include the
    shared header instead.

The shared helpers live in a named namespace (analyzer_util::) rather than at
global scope: the originals had internal linkage in anonymous namespaces, and
`candidates`/`line_of` are collision-prone generic names that should not land
on the global symbol table for every TU that includes the header. Each fish
TU pulls in exactly the two names it needs via `using` declarations.

The row/column switch now lives in exactly one place for both techniques.
Behavior is byte-for-byte identical: verbose [XW]/[fXW] and [SF]/[fSF] output
diffs clean against main on the X-Wing (P_clm) and both Swordfish paths. The
rewritten act_on_xwing is already covered on both branches by the existing
test_xwing_row_based / test_xwing_column_based whitebox tests.
